BirthdayCalendar creating notification service abstract class and renaming previous notification service class
